RTT-382: fixing memory leak: diffReports were being left in memory when no longer needed.

+++ rest/src/main/java/org/ihtsdo/otf/refset/rest/impl/RefsetServiceRestImpl.java
@@ -3222,8 +3222,10 @@ public class RefsetServiceRestImpl extends RootServiceRestImpl
       authorizeApp(securityService, authToken, "releases a report",
           UserRole.VIEWER);
 
-      refsetService.removeMembersInCommon(reportToken);
-      refsetService.removeMemberDiffReport(reportToken);
+      if(reportToken != null) {
+        refsetService.removeMembersInCommon(reportToken);
+        refsetService.removeMemberDiffReport(reportToken);
+      }
     } catch (Exception e) {
       handleException(e, "trying to release a report");
     } finally {
